# app/Http/Controllers/UserController.php
<?php

namespace App\Http\Controllers;

use App\Models\User;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Auth;
use Illuminate\Support\Facades\Storage;
use Illuminate\Support\Facades\Validator;
/**
 * @OA\Tag(
 *     name="User",
 *     description="API Endpoints for managing user details"
 * )
 */
class UserController extends Controller
{
    /**
     * @OA\Get(
     *     path="/api/user/get",
     *     summary="Get user details by ID",
     *     tags={"User"},
     *     @OA\Parameter(
     *         name="user_id",
     *         in="query",
     *         required=true,
     *         @OA\Schema(type="string")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="User details retrieved successfully",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="boolean", example=true),
     *             @OA\Property(property="message", type="string", example="User details retrieved successfully"),
     *             @OA\Property(
     *                 property="data",
     *                 type="object",
     *                 @OA\Property(property="id", type="string", example="string"),
     *                 @OA\Property(property="name", type="string", example="John Doe"),
     *                 @OA\Property(property="email", type="string", example="john@example.com"),
     *                 @OA\Property(property="image", type="string", example="userimage/user1.jpg"),
     *                 @OA\Property(property="email_verified_at", type="string", format="date-time", nullable=true),
     *                 @OA\Property(property="created_at", type="string", format="date-time"),
     *                 @OA\Property(property="updated_at", type="string", format="date-time")
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="User not found",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="User not found")
     *         )
     *     ),
     *     @OA\Response(
     *         response=400,
     *         description="Validation error",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="Validation error"),
     *             @OA\Property(property="errors", type="object")
     *         )
     *     )
     * )
     */
    public function getUser(Request $request)
    {
        $validator = Validator::make($request->all(), [
            'user_id' => 'required|string|size:36|exists:users,id',
        ]);

        if ($validator->fails()) {
            return response()->json([
                'status' => false,
                'message' => 'Validation error',
                'errors' => $validator->errors()
            ], 400);
        }

        $user = User::find($request->user_id);

        if (!$user) {
            return response()->json([
                'status' => false,
                'message' => 'User not found'
            ], 404);
        }

        return response()->json([
            'status' => true,
            'message' => 'User details retrieved successfully',
            'data' => $user
        ], 200);
    }

     /**
 * @OA\Post(
 *     path="/api/user/update",
 *     summary="Update user details by ID",
 *     tags={"User"},
 *     @OA\RequestBody(
 *         required=true,
 *         @OA\MediaType(
 *             mediaType="multipart/form-data",
 *             @OA\Schema(
 *                 @OA\Property(property="user_id", type="string", example="string"),
 *                 @OA\Property(property="name", type="string", example="John Doe"),
 *                 @OA\Property(property="email", type="string", format="email", example="john@example.com"),
 *                 @OA\Property(property="image", type="file", format="binary"),
 *                 @OA\Property(property="vat", type="string", example="GB123456789"),
 *                 @OA\Property(property="tin", type="string", example="123456789")
 *             )
 *         )
 *     ),
 *     @OA\Response(
 *         response=200,
 *         description="User updated successfully",
 *         @OA\JsonContent(
 *             @OA\Property(property="status", type="boolean", example=true),
 *             @OA\Property(property="message", type="string", example="User updated successfully"),
 *             @OA\Property(
 *                 property="data",
 *                 type="object",
 *                 @OA\Property(property="id", type="integer", example=1),
 *                 @OA\Property(property="name", type="string", example="John Doe"),
 *                 @OA\Property(property="email", type="string", example="john@example.com"),
 *                 @OA\Property(property="image", type="string", example="userimage/user1.jpg"),
 *                 @OA\Property(property="customer_type", type="string", example="company"),
 *                 @OA\Property(property="vat", type="string", example="GB123456789", nullable=true),
 *                 @OA\Property(property="tin", type="string", example="123456789", nullable=true),
 *                 @OA\Property(property="email_verified_at", type="string", format="date-time", nullable=true),
 *                 @OA\Property(property="created_at", type="string", format="date-time"),
 *                 @OA\Property(property="updated_at", type="string", format="date-time")
 *             )
 *         )
 *     ),
 *     @OA\Response(
 *         response=400,
 *         description="Validation error",
 *         @OA\JsonContent(
 *             @OA\Property(property="status", type="boolean", example=false),
 *             @OA\Property(property="message", type="string", example="Validation error"),
 *             @OA\Property(property="errors", type="object")
 *         )
 *     ),
 *     @OA\Response(
 *         response=404,
 *         description="User not found",
 *         @OA\JsonContent(
 *             @OA\Property(property="status", type="boolean", example=false),
 *             @OA\Property(property="message", type="string", example="User not found")
 *         )
 *     ),
 *     @OA\Response(
 *         response=403,
 *         description="Operation not allowed",
 *         @OA\JsonContent(
 *             @OA\Property(property="status", type="boolean", example=false),
 *             @OA\Property(property="message", type="string", example="VAT and TIN updates are only allowed for company or employee users")
 *         )
 *     )
 * )
 */
public function updateUser(Request $request)
{
    // First validate just the user_id to find the user
    $idValidator = Validator::make($request->all(), [
        'user_id' => 'required|string|size:36|exists:users,id',
    ]);

    if ($idValidator->fails()) {
        return response()->json([
            'status' => false,
            'message' => 'Validation error',
            'errors' => $idValidator->errors()
        ], 400);
    }

    $user = User::find($request->user_id);

    if (!$user) {
        return response()->json([
            'status' => false,
            'message' => 'User not found'
        ], 404);
    }

    // Create validation rules based on what's included in the request
    $rules = [];
    
    if ($request->filled('name')) {
        $rules['name'] = 'string|max:255';
    }
    
    if ($request->filled('email')) {
        $rules['email'] = 'string|email|max:255|unique:users,email,' . $user->id;
    }
    
    if ($request->hasFile('image')) {
        $rules['image'] = 'image|mimes:jpeg,png,jpg,gif|max:2048';
    }
    
    // Check if VAT or TIN fields are being updated
    $hasVatOrTin = $request->filled('vat') || $request->filled('tin');
    
    // If user is trying to update VAT or TIN but they're an individual customer type
    if ($hasVatOrTin && $user->customer_type === 'individual') {
        return response()->json([
            'status' => false,
            'message' => 'VAT and TIN updates are only allowed for company or employee users'
        ], 403);
    }
    
    // Add VAT and TIN validation rules if the user is company or employee
    if ($user->customer_type === 'company' || $user->customer_type === 'employee') {
        if ($request->filled('vat')) {
            $rules['vat'] = 'string|max:50';
        }
        
        if ($request->filled('tin')) {
            $rules['tin'] = 'string|max:50';
        }
    }
    
    // If there are any fields to validate, run the validation
    if (!empty($rules)) {
        $validator = Validator::make($request->all(), $rules);
        
        if ($validator->fails()) {
            return response()->json([
                'status' => false,
                'message' => 'Validation error',
                'errors' => $validator->errors()
            ], 400);
        }
    }

    // Update only the fields that were provided
    if ($request->filled('name')) {
        $user->name = $request->name;
    }
    
    if ($request->filled('email')) {
        $user->email = $request->email;
    }
    
    // Update VAT and TIN only if user is company or employee
    if ($user->customer_type === 'company' || $user->customer_type === 'employee') {
        if ($request->filled('vat')) {
            $user->vat = $request->vat;
        }
        
        if ($request->filled('tin')) {
            $user->tin = $request->tin;
        }
    }
    
    if ($request->hasFile('image')) {
        // Define the default image path
        $defaultImagePath = 'userimage/default.png';
        
        // Delete old image if it exists and is not the default image
        if ($user->image && 
            $user->image !== $defaultImagePath && 
            file_exists(public_path($user->image))) {
            unlink(public_path($user->image));
        }
        
        // Upload new image
        $image = $request->file('image');
        $imageName = time() . '_' . $user->id . '.' . $image->getClientOriginalExtension();
        $image->move(public_path('userimage'), $imageName);
        $user->image = 'userimage/' . $imageName;
    }
    
    $user->save();
    
    return response()->json([
        'status' => true,
        'message' => 'User updated successfully',
        'data' => $user
    ], 200);
}

public function getSubscribedUsersCount(Request $request)
{
    try {
        $count = User::where('email_subscribed', true)->count();

        return response()->json([
            'success' => true,
            'data' => ['count' => $count],
        ], 200);
    } catch (\Exception $e) {
        \Log::error('Failed to fetch subscribed users count: ' . $e->getMessage());
        return response()->json([
            'success' => false,
            'message' => 'Failed to fetch subscribed users count.',
        ], 500);
    }
}
}